# ARROW-5760: [C++] New compute::Take implementation for better performance, faster dispatch, smaller code size / faster compilation