--- a/cpp/src/arrow/compute/benchmark_util.h
+++ b/cpp/src/arrow/compute/benchmark_util.h
@@ -56,9 +56,10 @@ void BenchmarkSetArgsWithSizes(benchmark::internal::Benchmark* bench,
                                const std::vector<int64_t>& sizes = kMemorySizes) {
   bench->Unit(benchmark::kMicrosecond);
 
+  // 0 is treated as "no nulls"
   for (const auto size : sizes) {
     for (const auto inverse_null_proportion :
-         std::vector<ArgsType>({10000, 100, 10, 2, 1})) {
+           std::vector<ArgsType>({10000, 100, 50, 10, 2, 1, 0})) {
       bench->Args({static_cast<ArgsType>(size), inverse_null_proportion});
     }
   }
@@ -80,12 +81,16 @@ struct RegressionArgs {
   const int64_t size;
 
   // proportion of nulls in generated arrays
-  const double null_proportion;
+  double null_proportion;
 
   explicit RegressionArgs(benchmark::State& state)
-      : size(state.range(0)),
-        null_proportion(std::min(1., 1. / static_cast<double>(state.range(1)))),
-        state_(state) {}
+      : size(state.range(0)), null_proportion(), state_(state) {
+    if (state.range(1) == 0) {
+      this->null_proportion = 0.0;
+    } else {
+      this->null_proportion = std::min(1., 1. / static_cast<double>(state.range(1)));
+    }
+  }
 
   ~RegressionArgs() {
     state_.counters["size"] = static_cast<double>(size);