--- a/cpp/src/arrow/compute/kernels/vector_selection_benchmark.cc
+++ b/cpp/src/arrow/compute/kernels/vector_selection_benchmark.cc
@@ -90,7 +90,7 @@ struct TakeBenchmark {
         args(state),
         rand(kSeed),
         indices_have_nulls(indices_have_nulls),
-        monotonic_indices(false) {}
+        monotonic_indices(monotonic_indices) {}
 
   void Int64() {
     const int64_t array_size = args.size / sizeof(int64_t);
@@ -123,10 +123,10 @@ struct TakeBenchmark {
   }
 
   void Bench(const std::shared_ptr<Array>& values) {
-    bool indices_null_proportion = indices_have_nulls ? args.null_proportion : 0;
+    double indices_null_proportion = indices_have_nulls ? args.null_proportion : 0;
     auto indices =
-        rand.Int32(static_cast<int32_t>(values->length()), 0,
-                   static_cast<int32_t>(values->length() - 1), indices_null_proportion);
+        rand.Int32(values->length(), 0, static_cast<int32_t>(values->length() - 1),
+                   indices_null_proportion);
 
     if (monotonic_indices) {
       auto arg_sorter = *SortToIndices(*indices);
@@ -277,12 +277,12 @@ void TakeSetArgs(benchmark::internal::Benchmark* bench) {
 
 BENCHMARK(TakeInt64RandomIndicesNoNulls)->Apply(TakeSetArgs);
 BENCHMARK(TakeInt64RandomIndicesWithNulls)->Apply(TakeSetArgs);
+BENCHMARK(TakeInt64MonotonicIndices)->Apply(TakeSetArgs);
 BENCHMARK(TakeFSLInt64RandomIndicesNoNulls)->Apply(TakeSetArgs);
 BENCHMARK(TakeFSLInt64RandomIndicesWithNulls)->Apply(TakeSetArgs);
+BENCHMARK(TakeFSLInt64MonotonicIndices)->Apply(TakeSetArgs);
 BENCHMARK(TakeStringRandomIndicesNoNulls)->Apply(TakeSetArgs);
 BENCHMARK(TakeStringRandomIndicesWithNulls)->Apply(TakeSetArgs);
-BENCHMARK(TakeInt64MonotonicIndices)->Apply(TakeSetArgs);
-BENCHMARK(TakeFSLInt64MonotonicIndices)->Apply(TakeSetArgs);
 BENCHMARK(TakeStringMonotonicIndices)->Apply(TakeSetArgs);
 
 }  // namespace compute