--- a/cpp/src/arrow/util/bitmap_reader.h
+++ b/cpp/src/arrow/util/bitmap_reader.h
@@ -20,6 +20,7 @@
 #include <cstdint>
 #include <cstring>
 
+#include "arrow/buffer.h"
 #include "arrow/util/bit_util.h"
 #include "arrow/util/macros.h"
 
@@ -64,6 +65,19 @@ class BitmapReader {
   uint8_t current_byte_;
   int64_t byte_offset_;
   int64_t bit_offset_;
+};
+
+/// \brief Index into a possibly non-existent bitmap
+struct OptionalBitIndexer {
+  const uint8_t* bitmap;
+  const int64_t offset;
+
+  explicit OptionalBitIndexer(const std::shared_ptr<Buffer>& buffer, int64_t offset = 0)
+      : bitmap(buffer == NULLPTR ? NULLPTR : buffer->data()), offset(offset) {}
+
+  bool operator[](int64_t i) const {
+    return bitmap == NULLPTR ? true : BitUtil::GetBit(bitmap, offset + i);
+  }
 };
 
 }  // namespace internal