… cpp/src/arrow/compute/kernels/vector_take_test.cc
… cpp/src/arrow/util/bit_block_counter_test.cc
… cpp/src/arrow/util/int_util.h
… cpp/src/arrow/util/int_util_test.cc
… cpp/src/arrow/util/macros.h
… python/pyarrow/_compute.pyx
… python/pyarrow/compute.py
… python/pyarrow/includes/libarrow.pxd
… python/pyarrow/tests/test_pandas.py